server: Users, Roles and Apps APIs
Create user using,

```
POST /api/v1/users
```

Now login using,

```
POST /api/v1/apps
```

Above API returns `token`, `id`(App ID) and `user_id`. Save this for
future operations. Attach the following headers with every requests.

* `X-User-ID` - User ID
* `Authorization` - Authorization in `Bearer <token>` format.

Now create Cluster or any other operation with these headers.

Example:

```
USER_ID=user@example.com                                               \
TOKEN=ac73aad74b1378b4447033e2ce0b06a24500cf9c10f442138f83722d4f690376 \
curl -i -XPOST http://localhost:3000/api/v1/clusters                   \
    -H "Content-Type: application/json"                                \
    -H "X-User-ID: ${USER_ID}"                                         \
    -H "Authorization: Bearer ${TOKEN}"                                \
    -d '{"name": "neo"}'
```

+++ server/src/node_routes.cr
@@ -1,12 +1,21 @@
 require "kemal"
 
 require "./db/*"
+require "./helpers"
 
 get "/api/v1/clusters/:cluster_id/nodes" do |env|
+  if !cluster_viewer?(env)
+    halt(env, status_code: 403, response: forbidden_response)
+  end
+
   MoanaDB.list_nodes(env.params.url["cluster_id"]).to_json
 end
 
 get "/api/v1/clusters/:cluster_id/nodes/:id" do |env|
+  if !cluster_viewer?(env)
+    halt(env, status_code: 403, response: forbidden_response)
+  end
+
   node = MoanaDB.get_node(env.params.url["id"])
   if node.nil?
     env.response.status_code = 400
@@ -17,6 +26,10 @@ get "/api/v1/clusters/:cluster_id/nodes/:id" do |env|
 end
 
 post "/api/v1/clusters/:cluster_id/nodes" do |env|
+  if !cluster_maintainer?(env)
+    halt(env, status_code: 403, response: forbidden_response)
+  end
+
   hostname = env.params.json["hostname"].as(String)
   endpoint = env.params.json["endpoint"].as(String)
 
@@ -25,6 +38,10 @@ post "/api/v1/clusters/:cluster_id/nodes" do |env|
 end
 
 put "/api/v1/clusters/:cluster_id/nodes/:id" do |env|
+  if !cluster_maintainer?(env)
+    halt(env, status_code: 403, response: forbidden_response)
+  end
+
   hostname = env.params.json["hostname"]?.as(String?)
   endpoint = env.params.json["endpoint"]?.as(String?)
 
@@ -32,6 +49,10 @@ put "/api/v1/clusters/:cluster_id/nodes/:id" do |env|
 end
 
 delete "/api/v1/clusters/:cluster_id/nodes/:id" do |env|
+  if !cluster_maintainer?(env)
+    halt(env, status_code: 403, response: forbidden_response)
+  end
+
   MoanaDB.delete_node(env.params.url["id"])
 
   env.response.status_code = 204
